nyanmyoaung/phase_one/requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
Django==3.1.4
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
idna==2.10
phonenumbers==8.12.9
Pillow==7.2.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
PyJWT==2.0.1
pytz==2020.1
qrcode==6.1
requests==2.25.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.2
typing==3.7.4.3
urllib3==1.26.2
